feat: right-click a tab in a pane's favicon strip, and a setting to hide the strip
Two subagent efforts, landing together because both rewire the same panel.

**The strip's tabs get their own right-click menu.** The strip exists so a pane's
tabs can be reached without going to the sidebar; without the menu it sent you back
there for pin, split, bookmark and close - exactly the trip it was meant to save.

That needed the menu hoisted out of rememberTabBarState into TabMenuState.kt,
because the panel drawing the strip has no TabBarState to borrow: in LEFT position
the window owns the one bar and the panel deliberately builds none. Copying ~150
lines of menu actions there would have been a second copy to keep in step by hand.
The dialogs travel WITH the items rather than being left behind - two of the
entries do nothing on their own but raise a dialog, and a caller that took the
items and forgot the dialogs would get bookmark actions that silently did nothing.

**Settings > Window Appearance > Pane Tab Strip** turns the strip off. Both
existing conditions survive the setting rather than being replaced by it: still
LEFT position only, still only once the window is split. In TOP each pane already
draws a full strip, and with one pane the vertical bar lists every tab by name, so
the setting subtracts from those cases and cannot add to them. Default on, because
that is what a split window has done since the strip existed.

Finished from where the agents stopped - both died mid-run on an API weekly limit.
What was left: a wildcard icon import, no settings-search entry for the new toggle
(SettingsSearchIndexDriftTest catches exactly that), and a file whose name did not
match its one class.

The detekt baseline gains two size entries. rememberTabBarState shed 274 lines and
is still over the limit, so what used to be one oversized function is now two - the
code did not grow, the entry count did.
