+++ composeApp/src/desktopMain/kotlin/ai/rever/boss/components/settings/search/SettingsSearchEntries.kt
@@ -301,6 +301,7 @@ private fun windowAppearanceEntries() =
         setting("Tab Sizing", "Tab Bar", "shrink to fit", "fixed width", "tabs")
         setting("Vertical Bar Width", "Tab Bar", "vertical tabs", "left", "sidebar", "width")
         setting("Expand on Hover", "Tab Bar", "vertical tabs", "rail", "collapse", "drawer")
+        setting("Pane Tab Strip", "Tab Bar", "vertical tabs", "split", "favicon", "pane", "tabs")
         group("Bars")
         setting("Show Top Bar", "Bars", "chrome", "window")
         setting("Show Bottom Bar", "Bars", "chrome", "window", "status")
